move expression->value logic into ValueBuilder
