+++ Source/Parser/Functions/RichPresenceConditionalDisplayFunction.cs
@@ -1,8 +1,4 @@
-﻿using RATools.Data;
-using RATools.Parser.Expressions;
-using RATools.Parser.Internal;
-using System.Collections.Generic;
-using System.Diagnostics;
+﻿using RATools.Parser.Expressions;
 
 namespace RATools.Parser.Functions
 {
